Add an index to the job_history table to be able to query it by trigger
Currently the job_history table has indices to query by job name and group.
The table also stores trigger information, but has no index to query it.
This adds an index to be able to query by trigger name and group, so the
ability to query job history by trigger added in 1.15 is not incredibly
slow.